Allow use of tags in description
If the tag list is not introduced by a comma, it's supposed to
be part of the description... so we copy the embedded tags to the tag
list and we drop the leading hash to keep the plain word in the
description.

Tags embedded in the description are more restricted in their syntax,
they can't contain spaces and must not start with a digit (we want to
avoid creating tags for things like bug numbers).

So this basically converts a fact like "Coding, fix #bugs in #hamster"
into "Coding, fix bugs in hamster, #bugs #hamster".